move get_lookup_dicts to wiki
